pyproject.toml
[tool.poetry]
name = "django-ozon-parser"
version = "0.1.0"
description = ""
authors = ["Rusih100 <rusih100@gmail.com>"]
readme = "README.md"
packages = [{include = "django_ozon_parser"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "3.1.14"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-mysql = "^3.8.1"
mysqlclient = "^2.2.0"
selenium = "^4.11.2"
celery = {extras = ["redis"], version = "^5.3.1"}


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
